fix(api): make groups.pool_memory_mb BIGINT to match the i64 entity
The 000025 migration created pool_memory_mb as INT but the sea-orm entity
decodes it as Option<i64> (INT8), so every group row read failed with a
ColumnDecode mismatch. BIGINT matches the other memory columns (template
memory, system_settings.host_memory_mb) and the i64 usage math.

+++ apps/api/migration/src/m20260809_000025_add_resource_quotas.rs
@@ -28,7 +28,7 @@ impl MigrationTrait for Migration {
             ALTER TABLE groups
                 ADD COLUMN billing_model VARCHAR(16) NOT NULL DEFAULT 'shared',
                 ADD COLUMN pool_cpu_cores INT NOT NULL DEFAULT 0,
-                ADD COLUMN pool_memory_mb INT NOT NULL DEFAULT 0,
+                ADD COLUMN pool_memory_mb BIGINT NOT NULL DEFAULT 0,
                 ADD COLUMN pool_gpu_count INT NOT NULL DEFAULT 0
             "#,
         )
